fix: Set GEM_PATH and RUBYLIB in snap environment
Turns out that Ruby didn't actually work inside the snap without having these paths set. It was piggybacking of the Ruby gems that I had installed on my devbox. Found the problem when `ood-portal-generator` failed to create the portal. Now, setting RUBYLIB and GEM_PATH to the right paths enables `ood-portal-generator` to work out of the box without requiring gems installed on the host system.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>