fix: dynamically load modules in httpd.conf
The lua module must be loaded after the
env module, or lua will be unable to
access the environment variables set by
`SetEnv` in the httpd configuration files.
I discovered the fix after manually
loading lua after env as a DSO. I don't
understand why this fixed the issue, but it
works so I'm not going to complain.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>